Send Batch Email Using a Template

Send email to more than 500 recipients, sending multiple POST requests via httr2::req_perform_sequential().

Usage

template_send_email_batch(
  from,
  to,
  id,
  template_model,
  msg_stream,
  tag = NULL,
  track_opens = FALSE,
  token = NULL
)

Arguments

from

Character scalar. Email address of the sender.

to

Character vector. Email addresses of recipients (max 50).

id

A single integer. The template ID in Postmark.

template_model

A named list. Variables to be populated in the template.

msg_stream

Character scalar. Either "outbound" or "broadcast".

tag

A single character string. Optional tag for categorizing the email. Maximum 1000 characters. Default is NULL.

track_opens

A logical value. Whether to track when recipients open the email. Default is FALSE.

token

Character string. Your Postmark API token. If NULL (default), the function will attempt to retrieve the token using get_token() POSTMARK_SERVER_TOKEN as an environment variable. Check the Postmark's API documentation on how to get a server token.

Value

A data frame containing the JSON response from the Postmark API, invisibly.